third_party/blink/web_tests/platform/generic/fast/dom/Element/id-in-node-list-index01-expected.txt

Test using id as index of node list

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS elems["id1"].getAttribute("name") is "name1"
PASS elems["id2"].getAttribute("name") is "name2"
PASS successfullyParsed is true

TEST COMPLETE
